Retirado BasicSelect e CitySelect de Adapters; Adicionado CitiesService.listByState(stateUid)
